third_party/blink/web_tests/external/wpt/svg/linking/scripted/href-animate-element-expected.txt

This is a testharness.js-based test.
[FAIL] Test for animate element when removing href but we still have xlink:href
  assert_equals: expected 100 but got 50
Harness: the test ran to completion.
